requirements-lite.txt
# Minimal dependencies for LGSP project (Python 3.9 compatible)
# Install PyTorch separately per your CUDA setup from https://pytorch.org/get-started/locally/

numpy>=1.21,<1.24
scipy>=1.7,<1.9
pandas>=1.3,<1.6
Pillow>=9.0,<10
matplotlib>=3.5,<3.8
scikit-learn==1.3.2
opencv-python==4.7.0.72

einops==0.6.1
kornia==0.7.0
yacs==0.1.8
easydict==1.10
pytorch-metric-learning==2.3.0
POT==0.9.0
geomloss==0.2.6
randaugment==1.0.2

# Vision/transformer tooling
ftfy==6.1.1
timm==0.6.12
transformers==4.27.4
huggingface-hub==0.12.1
safetensors==0.3.1
tokenizers==0.13.3
sentencepiece==0.1.99
open-clip-torch==2.20.0

# Utilities
requests>=2.27,<2.29
tqdm>=4.64,<5
